# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ESD Ticket Booking Platform API
  description: |
    Microservices-based event ticket booking platform for IS213.

    All responses follow a standard envelope:
    - **Success**: `{"code": N, "data": {...}}`
    - **Error**: `{"code": N, "message": "..."}`

    All endpoints accept an optional `X-Correlation-ID` header (UUID). If not provided, one is auto-generated for distributed tracing.
  version: 1.0.0
  contact:
    name: ESD Team
servers:
  - url: http://localhost:8000
    description: Kong API Gateway (local development)
...
paths:
...
  /api/tickets/booking/{booking_id}:
    get:
      tags: [Tickets]
      summary: Get ticket by booking ID
      description: Returns the ticket for a booking, including a base64-encoded QR code image.
      operationId: getTicketByBooking
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer
          description: Booking ID
      responses:
        "200":
          description: Ticket with QR code
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tickets/booking/{booking_id}/invalidate:
    post:
      tags: [Tickets]
      summary: Invalidate a ticket
      description: Marks a ticket as invalidated (used during refunds).
      operationId: invalidateTicket
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer
          description: Booking ID
      responses:
        "200":
          description: Ticket invalidated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tickets/{ticket_id}:
    get:
      tags: [Tickets]
      summary: Get ticket by ticket ID
      operationId: getTicketById
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
        - name: ticket_id
          in: path
          required: true
          schema:
            type: integer
          description: Ticket ID
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tickets/validate:
    post:
      tags: [Tickets]
      summary: Validate ticket QR code
      description: |
        Validates a scanned QR code at event entry. Verifies the HMAC hash,
        checks ticket status, and marks it as "used". Returns 409 if already
        used, 410 if invalidated (refunded/cancelled), 403 if hash mismatch.

        QR data format: `{booking_id}:{sha256_hash_first_16_chars}`

        Used by the admin QR scanner page (`/scanner`). The frontend enriches
        the response with event name and seat details from the Event and Seat services.
      operationId: validateTicket
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qr_data]
              properties:
                qr_data:
                  type: string
                  description: Raw QR code data string (format booking_id:hash)
                  example: "2:a1b2c3d4e5f6g7h8"
      responses:
        "200":
          description: Ticket validated and marked as used
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ticket_id:
                            type: integer
                          booking_id:
                            type: integer
                          event_id:
                            type: integer
                          user_id:
                            type: string
                          seat_id:
                            type: integer
                          status:
                            type: string
                            example: used
                          message:
                            type: string
                            example: Ticket validated successfully
        "400":
          description: Invalid QR code format
        "403":
          description: QR code hash mismatch
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Ticket already used
        "410":
          description: Ticket invalidated (refunded or event cancelled)
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────────────
  # Waitlist
  # ─────────────────────────────────────────────
...
components:
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Correlation ID for distributed tracing. Auto-generated if not provided.
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"

    # ── Event ───────────────────────────────────
...
    Ticket:
      type: object
      properties:
        ticket_id:
          type: integer
          example: 77
        booking_id:
          type: integer
          example: 101
        event_id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "firebase_uid_abc123"
        seat_id:
          type: integer
          example: 42
        qr_code_data:
          type: string
          description: Raw data encoded in the QR code
          example: "TICKET-77-BOOKING-101"
        status:
          type: string
          enum: [valid, invalidated]
          example: "valid"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        qr_code_base64:
          type: string
          description: Base64-encoded PNG image of the QR code
          example: "iVBORw0KGgoAAAANSUhEUgAA..."

    # ── Waitlist ────────────────────────────────
...